// Dashboard/src/Types/Tables.type.ts
export interface Permission {
    _id: string;
    permission: string;
    shortcode: string;

}